% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.dim.b.R
\name{ds.dim.b}
\alias{ds.dim.b}
\title{Retrieves the dimension of an object}
\usage{
ds.dim.b(x = NULL, type = "both", datasources = NULL)
}
\arguments{
\item{x}{a character, the name of R table object, for example a matrix, array or data frame}

\item{type}{a character which represents the type of analysis to carry out. 
If \code{type} is set to 'combine', a the combined dimensions of the object from all the
studies is returned. If \code{type} is set to 'split', the dimensions of the object from
each single study is returned. If \code{type} is set to 'both' (default), the dimensions 
of the object from each single study and the dimensions of the object in the combined case 
are returned.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
for an array, \code{NULL} or a vector of mode \code{integer}
}
\description{
this function is similar to R function \code{dim}
}
\details{
the function returns the unpooled or pooled dimension of the object by summing
up the individual dimensions returned from each study or the dimension of the object in each 
study. Unlike the other DataSHIELD function the default behaviour is to output the dimension 
of each study separately.
}
\examples{
{

  # load that contains the login details
  data(logindata)

  # login and assign all the stored variables.
  opals <- datashield.login(logins=logindata,assign=TRUE)

  # Example 1: Get the dimension of the assigned datasets in each study
  ds.dim(x='D', type='combine')

  # Example 2: Get the pooled dimension of the assigned datasets
  ds.dim(x='D', type='combine')

  # Example 3: Get the dimension og the datasets in each single study
  # and the pooled dimension  - default
  ds.dim(x='D') 

  # Example 4: Input has to be either matrix, data frame or an array
  # In the below example, the inpout is a vector so it will not work.
  \dontrun{ ds.dim(x='D$LAB_TSC') }

  # clear the Datashield R sessions and logout
  datashield.logout(opals)

}

}
\seealso{
\link{ds.dataframe} to generate a table of type dataframe.

\link{ds.changeRefGroup} to change the reference level of a factor.

\link{ds.colnames} to obtain the column names of a matrix or a data frame

\link{ds.asMatrix} to coerce an object into a matrix type.

\link{ds.length} to obtain the size of a vector.
}
\author{
Gaye, A.; Isaeva, J.; Avraam, D.
}
