% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.glm.REMA.R
\name{ds.glm.REMA}
\alias{ds.glm.REMA}
\title{ds.glm.REMA}
\usage{
ds.glm.REMA(formula = NULL, data = NULL, family = NULL, offset = NULL,
  weights = NULL, checks = FALSE, maxit = 15, CI = 0.95,
  viewIter = FALSE, datasources = NULL)
}
\description{
TODO
}

