% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.glmerSLMA.o.R
\name{ds.glmerSLMA.o}
\alias{ds.glmerSLMA.o}
\title{ds.glmSLMA.o calling glmDS1.o, glmSLMADS2.o}
\usage{
ds.glmerSLMA.o(formula = NULL, offset = NULL, weights = NULL, family,
  combine.with.metafor = TRUE, dataName = NULL, checks = FALSE,
  datasources = NULL, nAGQ = 1)
}
\description{
Fits a generalized linear model (glm) on data from a single or multiple sources
}
\details{
Fits a thing
}
