% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.glmSLMA.o.R
\name{ds.glmSLMA}
\alias{ds.glmSLMA}
\title{@title ds.glm calling glmDS1, glmDS2}
\usage{
ds.glmSLMA(formula = NULL, family = NULL, offset = NULL,
  weights = NULL, dataName = NULL, checks = FALSE, maxit = 15,
  CI = 0.95, viewIter = FALSE, viewVarCov = FALSE, viewCor = FALSE,
  datasources = NULL)
}
\description{
@title ds.glm calling glmDS1, glmDS2
}
