% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.length.b.R
\name{ds.length.b}
\alias{ds.length.b}
\title{Gets the length of a vector or list}
\usage{
ds.length.b(x = NULL, type = "both", datasources = NULL)
}
\arguments{
\item{x}{a string character, the name of a vector}

\item{type}{a character which represents the type of analysis to carry out. 
If \code{type} is set to 'combine', a the combined length of the vector from all the
studies is returned. If \code{type} is set to 'split', the length of the vector from
each single study is returned. If \code{type} is set to 'both' (default), the length 
of the vector from each single study and the length of the vector in the combined case 
are returned.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
a numeric, the number of elements of the input vector or list.
}
\description{
This function is similar to R function \code{length}.
}
\details{
The function returns the pooled length or the length of the a vector or a list for 
each study.
}
\examples{
{

  # load that contains the login details
  data(logindata)

  # login and assign all the variables stored on the server side
  opals <- datashield.login(logins=logindata,assign=TRUE)

  # Example 1: Get the total number of observations across all the studies for the variable 'LAB_TSC'
  ds.length(x='D$LAB_TSC', type='combine')

  # Example 2: Get the number of observations on each study, for the variable 'LAB_TSC'
  ds.length(x='D$LAB_TSC', type='split')

  # Example 3: Get the number of observations on each study and the total number of observations
  # across all the studies for the variable 'LAB_TSC' - default behaviour 

  # clear the Datashield R sessions and logout
  datashield.logout(opals) 

}

}
\author{
Gaye, A.; Avraam, D.
}
