% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.lmerSLMA.o.R
\name{ds.lmerSLMA.o}
\alias{ds.lmerSLMA.o}
\title{ds.glmSLMA.o calling glmDS1.o, calling glmDS1.o, glmSLMADS2.o}
\usage{
ds.lmerSLMA.o(formula = NULL, offset = NULL, weights = NULL,
  combine.with.metafor = TRUE, dataName = NULL, checks = FALSE,
  maxit = 15, datasources = NULL)
}
\description{
Fits a linear mixed effects model (lme) on data from a single or multiple sources
}
\details{
Fits a linear mixed effects model (lme) on data from a single or multiple sources
}
\author{
Tom Bishop
}
