% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.lmeSLMA.o.R
\name{ds.lmeSLMA.o}
\alias{ds.lmeSLMA.o}
\title{ds.glmSLMA.o calling glmDS1.o, glmSLMADS2.o}
\usage{
ds.lmeSLMA.o(formula = NULL, offset = NULL, weights = NULL,
  combine.with.metafor = TRUE, dataName = NULL, checks = FALSE,
  maxit = 15, datasources = NULL)
}
\description{
Fits a generalized linear model (glm) on data from a single or multiple sources
}
\details{
Fits a thing
}
