% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.rilm.b.R
\name{ds.rilm.b}
\alias{ds.rilm.b}
\title{Random Intercept Linear Model (RILM)}
\usage{
ds.rilm.b(..., y = NULL, datasources = NULL)
}
\arguments{
\item{...}{the explanatory variables, could be factor or numeric vectors}

\item{y}{the dependent variable}

\item{datasources}{a list of opal object(s) obtained after login to opal servers;
these objects also hold the data assigned to R, as a \code{dataframe}, from opal datasources.}
}
\value{
regression outcome
}
\description{
Runs a Random Intercept Linear Model on horizontally-partinioned data
}
\details{
see description
}
\author{
Avraam D; Jones EM; van den Heuvel ER
}
