% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tTestF.R
\name{ds.tTestF}
\alias{ds.tTestF}
\title{Runs a combined GLM analysis of non-pooled data}
\usage{
ds.tTestF(formula = NULL, data = NULL, family = "gaussian",
  offset = NULL, weights = NULL, checks = FALSE, maxit = 15,
  CI = 0.95, viewIter = FALSE, datasources = NULL)
}
\arguments{
\item{formula}{a character, a formula which describes the model to be fitted}

\item{data}{a character, the name of an optional data frame containing the variables in the
\code{formula}. The process stops if a non existing data frame is indicated.}

\item{family}{a description of the error distribution function to use in the model}

\item{offset}{a character, null or a numeric vector that can be used to specify an a priori
known component to be included in the linear predictor during fitting.}

\item{weights}{a character, the name of an optional vector of 'prior weights' to be used in the
fitting process. Should be NULL or a numeric vector.}

\item{checks}{a boolean, if TRUE (default) checks that takes 1-3min are carried out to verify
that the variables in the model are defined (exist) on the server site and that they have the
correct characteristics required to fit a GLM. The default value is FALSE because checks
lengthen the runtime and are mainly meant to be used as help to look for causes of eventual
errors.}

\item{maxit}{the number of iterations of IWLS used
instructions to each computer requesting non-disclosing summary statistics.
The summaries are then combined to estimate the parameters of the model; these
parameters are the same as those obtained if the data were 'physically' pooled.}

\item{CI}{a numeric, the confidence interval.}

\item{viewIter}{a boolean, tells whether the results of the intermediate iterations
should be printed on screen or not. Default is FALSE (i.e. only final results are shown).}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login. If the <datasources>
argument is not specified the default set of connections will be used: see \link{datashield.connections_default}.}
}
\value{
coefficients a named vector of coefficients

residuals the 'working' residuals, that is the residuals in the final
iteration of the IWLS fit.

fitted.values the fitted mean values, obtained by transforming the
linear predictors by the inverse of the link function.

rank the numeric rank of the fitted linear model.

family the \code{family} object used.

linear.predictors the linear fit on link scale.
}
\description{
A function fit generalized linear models
}
\details{
It enables a parallelized analysis of individual-level data sitting
on distinct servers by sending
}
\examples{
\donttest{
# # load the file that contains the login details
# logindata <- DSLite::setupCNSIMTest("dsBetaTest")
#
# # login and assign all the variables to R
# conns <- datashield.login(logins=logindata, assign=TRUE)
#
# # Example 1: run a GLM without interaction (e.g. diabetes prediction using BMI and HDL levels
# # and GENDER)
# mod <- ds.glm(formula='D$DIS_DIAB~D$GENDER+D$PM_BMI_CONTINUOUS+D$LAB_HDL', family='binomial')
# mod
#
# # Example 2: run the above GLM model without an intercept
# # (produces separate baseline estimates for Male and Female)
# mod <- ds.glm(formula='D$DIS_DIAB~0+D$GENDER+D$PM_BMI_CONTINUOUS+D$LAB_HDL', family='binomial')
# mod
#
# # Example 3: run the above GLM with interaction between GENDER and PM_BMI_CONTINUOUS
# mod <- ds.glm(formula='D$DIS_DIAB~D$GENDER*D$PM_BMI_CONTINUOUS+D$LAB_HDL', family='binomial')
# mod
#
# # Example 4: Fit a standard Gaussian linear model with an interaction
# mod <- ds.glm(formula='D$PM_BMI_CONTINUOUS~D$DIS_DIAB*D$GENDER+D$LAB_HDL', family='gaussian')
# mod
#
# # Example 5: now run a GLM where the error follows a poisson distribution
# # P.S: A poisson model requires a numeric vector as outcome so in this example we first convert
# # the categorical BMI, which is of type 'factor', into a numeric vector
# ds.asNumeric('D$PM_BMI_CATEGORICAL','BMI.123')
# mod <- ds.glm(formula='BMI.123~D$PM_BMI_CONTINUOUS+D$LAB_HDL+D$GENDER', family='poisson')
# mod
#
# # clear the Datashield R sessions and logout
# datashield.logout(conns)
}

}
\seealso{
\link{ds.lexis} for survival analysis using piecewise exponential regression

\link{ds.gee} for generalized estimating equation models
}
\author{
Burton PR; Gaye A; LaFlamme P
}
