% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTestHelper1.R
\name{tTestHelper1}
\alias{tTestHelper1}
\title{runs a t-test for two continuous variables}
\usage{
tTestHelper1(x, y, type, alternative, mu, paired, var.equal, conf.level,
  datasources)
}
\arguments{
\item{type}{see the calling function}

\item{alternative}{see the calling function}

\item{mu}{see the calling function}

\item{paired}{see the calling function}

\item{var.equal}{see the calling function}

\item{conf.level}{see the calling function}

\item{datasources}{see the calling function}
}
\value{
the results of the t-test
}
\description{
This is an internal function.
}
\details{
internal function
}
\keyword{calling}
\keyword{function}
\keyword{internal}
\keyword{see}
\keyword{the}
