% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTestHelper2.R
\name{tTestHelper2}
\alias{tTestHelper2}
\title{Uses glm to compute means of numeric vector across factor vector}
\usage{
tTestHelper2(formula, CI, datasources)
}
\arguments{
\item{formula}{a character form of a formula (e.g. 'a~b')}

\item{CI}{a numeric, the confidence interval}

\item{datasources}{see the calling function}

\item{family}{a description of the error distribution function to use in the model}
}
\value{
results similar to that of a t.test
}
\description{
This is an internal function.
}
\details{
internal function
}
\keyword{calling}
\keyword{function}
\keyword{internal}
\keyword{see}
\keyword{the}
